Move regexp_error_msg to utils/misc.py
